+++ src/test/java/ca/corefacility/bioinformatics/irida/service/impl/integration/ProjectServiceImplIT.java
@@ -750,14 +750,14 @@ public class ProjectServiceImplIT {
 
 	@Test
 	@WithMockUser(username = "user1", roles = "USER")
-	public void testMoveSamplesWithOwner() {
+	public void testMoveSamples() {
 		Project source = projectService.read(2L);
 		Project destination = projectService.read(10L);
 
 		Sample sample1 = sampleService.read(1L);
 		Set<Sample> samples = Sets.newHashSet(sample1);
 
-		List<ProjectSampleJoin> movedSamples = projectService.moveSamples(source, destination, samples, true);
+		List<ProjectSampleJoin> movedSamples = projectService.moveSamples(source, destination, samples);
 
 		assertEquals(samples.size(), movedSamples.size());
 
@@ -788,7 +788,7 @@ public class ProjectServiceImplIT {
 
 	@Test(expected = AccessDeniedException.class)
 	@WithMockUser(username = "user1", roles = "USER")
-	public void testMoveSamplesWithOwnerRemoteFail() {
+	public void testMoveSamplesRemoteFail() {
 		Project source = projectService.read(11L);
 		Project destination = projectService.read(10L);
 
@@ -798,7 +798,7 @@ public class ProjectServiceImplIT {
 		Sample sample = sampleService.read(3L);
 		Set<Sample> samples = Sets.newHashSet(sample);
 
-		projectService.moveSamples(source, destination, samples, true);
+		projectService.moveSamples(source, destination, samples);
 	}
 
 	@Test
@@ -827,30 +827,6 @@ public class ProjectServiceImplIT {
 						.collect(Collectors.toSet()));
 	}
 
-	@Test
-	@WithMockUser(username = "user1", roles = "USER")
-	public void testMoveSamplesWithoutOwner() {
-		Project source = projectService.read(2L);
-		Project destination = projectService.read(10L);
-
-		List<Join<Project, Sample>> samplesForProject = sampleService.getSamplesForProject(source);
-
-		Set<Sample> samples = samplesForProject.stream().map(j -> j.getObject()).collect(Collectors.toSet());
-
-		List<ProjectSampleJoin> movedSamples = projectService.moveSamples(source, destination, samples, false);
-
-		assertEquals(samples.size(), movedSamples.size());
-
-		movedSamples.forEach(j -> {
-			assertFalse("Project shouldn't be owner for sample", j.isOwner());
-		});
-
-		assertEquals("Samples should not exist in source project", Long.valueOf(0L),
-				projectSampleJoinRepository.countSamplesForProject(source));
-		assertEquals("Samples should exist in destination project", Sets.newHashSet(1L, 2L),
-				projectSampleJoinRepository.getSamplesForProject(destination).stream().map(j -> j.getObject().getId())
-						.collect(Collectors.toSet()));
-	}
 
 	@Test
 	@WithMockUser(username = "user1", roles = "USER")
@@ -876,21 +852,6 @@ public class ProjectServiceImplIT {
 				projectSampleJoinRepository.readSampleForProject(source, sample));
 		assertNotNull("Sample should exist in destination project",
 				projectSampleJoinRepository.readSampleForProject(destination, sample));
-	}
-
-	@Test(expected = AccessDeniedException.class)
-	@WithMockUser(username = "user1", roles = "USER")
-	public void testMoveSamplesWithoutOwnerRemoteFail() {
-		Project source = projectService.read(11L);
-		Project destination = projectService.read(10L);
-
-		assertTrue("Source project should be a remote project for the test", source.isRemote());
-		assertFalse("Destination project should not be a remote project for the test", destination.isRemote());
-
-		Sample sample = sampleService.read(3L);
-		Set<Sample> samples = Sets.newHashSet(sample);
-
-		projectService.moveSamples(source, destination, samples, false);
 	}
 
 	@Test
@@ -925,11 +886,11 @@ public class ProjectServiceImplIT {
 		Project source = projectService.read(2L);
 		Project destination = projectService.read(10L);
 
-		List<Join<Project, Sample>> samplesForProject = sampleService.getSamplesForProject(source);
+		Sample sample2 = sampleService.read(2L);
 
-		Set<Sample> samples = samplesForProject.stream().map(j -> j.getObject()).collect(Collectors.toSet());
+		Set<Sample> samples = Sets.newHashSet(sample2);
 
-		List<ProjectSampleJoin> movedSamples = projectService.moveSamples(source, destination, samples, false);
+		List<ProjectSampleJoin> movedSamples = projectService.moveSamples(source, destination, samples);
 
 		assertEquals(samples.size(), movedSamples.size());
 
@@ -937,11 +898,14 @@ public class ProjectServiceImplIT {
 			assertFalse("Project shouldn't be owner for sample", j.isOwner());
 		});
 
-		assertEquals("Samples should not exist in source project", Long.valueOf(0L),
+		assertEquals("Samples should not exist in source project", Long.valueOf(1L),
 				projectSampleJoinRepository.countSamplesForProject(source));
-		assertEquals("Samples should exist in destination project", Sets.newHashSet(1L, 2L),
-				projectSampleJoinRepository.getSamplesForProject(destination).stream().map(j -> j.getObject().getId())
-						.collect(Collectors.toSet()));
+
+		List<Join<Project, Sample>> samplesForProject = projectSampleJoinRepository.getSamplesForProject(destination);
+		assertEquals("Should be 1 sample", 1, samplesForProject.size());
+		ProjectSampleJoin join = (ProjectSampleJoin) samplesForProject.iterator().next();
+		assertFalse("Project should not be owner", join.isOwner());
+
 	}
 
 	@Test(expected = AccessDeniedException.class)
@@ -955,19 +919,6 @@ public class ProjectServiceImplIT {
 		Set<Sample> samples = samplesForProject.stream().map(j -> j.getObject()).collect(Collectors.toSet());
 
 		projectService.shareSamples(source, destination, samples, true);
-	}
-
-	@Test(expected = AccessDeniedException.class)
-	@WithMockUser(username = "user1", roles = "USER")
-	public void testMoveLockedSamplesWithOwnerFail() {
-		Project source = projectService.read(2L);
-		Project destination = projectService.read(10L);
-
-		List<Join<Project, Sample>> samplesForProject = sampleService.getSamplesForProject(source);
-
-		Set<Sample> samples = samplesForProject.stream().map(j -> j.getObject()).collect(Collectors.toSet());
-
-		projectService.moveSamples(source, destination, samples, true);
 	}
 
 	private Project p() {